refactor(main.py): main.py is now in the root folder. main.py has been exploded in multiple subpart
cli_centraliser centraliser cli.add_command ; job_scrapper now contain a `cli` folder that contains cli for 'scrapping' and 'database request' #35
