+++ cli/cli_centraliser.py
@@ -1,0 +1,4 @@
+from job_scrapper.cli.scrapper_cli import scrap
+from cli.cli_base import cli
+
+cli.add_command(scrap)
