+++ main.py
@@ -1,0 +1,9 @@
+from cli.cli_centraliser import cli
+
+def main():
+    """Start Application's Command Line Interface"""
+    cli()
+
+
+if __name__ == "__main__":
+    cli()
