+++ pyproject.toml
@@ -8,7 +8,7 @@ license = "CC BY-SA 4.0"
 packages = [{ include = "job_scrapper" }]
 
 [tool.poetry.scripts]
-job-scrapper = "job_scrapper.main:cli"
+job-scrapper = "main:cli"
 
 [tool.poetry.dependencies]
 python = ">=3.10.12,<3.11"
